add ABINet++

+++ adet/evaluation/text_evaluation.py
@@ -21,6 +21,7 @@ import shutil
 from shapely.geometry import Polygon, LinearRing
 from adet.evaluation import text_eval_script
 import zipfile
+import pickle
 
 
 class TextEvaluator(DatasetEvaluator):
@@ -41,6 +42,15 @@ class TextEvaluator(DatasetEvaluator):
             raise AttributeError(
                 f"json_file was not found in MetaDataCatalog for '{dataset_name}'."
             )
+        
+        self.voc_size = cfg.MODEL.BATEXT.VOC_SIZE
+        self.use_customer_dictionary = cfg.MODEL.BATEXT.CUSTOM_DICT
+        if not self.use_customer_dictionary:
+            self.CTLABELS = [' ','!','"','#','$','%','&','\'','(',')','*','+',',','-','.','/','0','1','2','3','4','5','6','7','8','9',':',';','<','=','>','?','@','A','B','C','D','E','F','G','H','I','J','K','L','M','N','O','P','Q','R','S','T','U','V','W','X','Y','Z','[','\\',']','^','_','`','a','b','c','d','e','f','g','h','i','j','k','l','m','n','o','p','q','r','s','t','u','v','w','x','y','z','{','|','}','~']
+        else:
+            with open(self.use_customer_dictionary, 'rb') as fp:
+                self.CTLABELS = pickle.load(fp)
+        assert(int(self.voc_size - 1) == len(self.CTLABELS)), "voc_size is not matched dictionary size, got {} and {}.".format(int(self.voc_size - 1), len(self.CTLABELS))
 
         json_file = PathManager.get_local_path(self._metadata.json_file)
         with contextlib.redirect_stdout(io.StringIO()):
@@ -50,9 +60,14 @@ class TextEvaluator(DatasetEvaluator):
         if "totaltext" in dataset_name:
             self._text_eval_gt_path = "datasets/evaluation/gt_totaltext.zip"
             self._word_spotting = True
+        elif "icdar2015" in dataset_name:
+            self._text_eval_gt_path = "datasets/evaluation/gt_icdar2015.zip"
+            self._word_spotting = False
         elif "ctw1500" in dataset_name:
             self._text_eval_gt_path = "datasets/evaluation/gt_ctw1500.zip"
             self._word_spotting = False
+        else:
+            self._text_eval_gt_path = ""
         self._text_eval_confidence = cfg.MODEL.FCOS.INFERENCE_TH_TEST
 
     def reset(self):
@@ -63,7 +78,7 @@ class TextEvaluator(DatasetEvaluator):
             prediction = {"image_id": input["image_id"]}
 
             instances = output["instances"].to(self._cpu_device)
-            prediction["instances"] = instances_to_coco_json(instances, input["image_id"])
+            prediction["instances"] = self.instances_to_coco_json(instances, input["image_id"])
             self._predictions.append(prediction)
 
     def to_eval_format(self, file_path, temp_dir="temp_det_results", cf_th=0.5):
@@ -80,7 +95,7 @@ class TextEvaluator(DatasetEvaluator):
 
         with open(file_path, 'r') as f:
             data = json.load(f)
-            with open('temp_all_det_cors.txt', 'w') as f2:
+            with open(os.path.join(temp_dir,'temp_all_det_cors.txt'), 'w') as f2:
                 for ix in range(len(data)):
                     if data[ix]['score'] > 0.1:
                         outstr = '{}: '.format(data[ix]['image_id'])
@@ -97,7 +112,7 @@ class TextEvaluator(DatasetEvaluator):
                 f2.close()
         dirn = temp_dir
         lsc = [cf_th] 
-        fres = open('temp_all_det_cors.txt', 'r').readlines()
+        fres = open(os.path.join(temp_dir,'temp_all_det_cors.txt'), 'r').readlines()
         for isc in lsc:	
             if not os.path.isdir(dirn):
                 os.mkdir(dirn)
@@ -114,16 +129,16 @@ class TextEvaluator(DatasetEvaluator):
                         continue
                     cors = ','.join(e for e in ptr[0].split(',')[:-1])
                     fout.writelines(cors+',####'+ptr[1]+'\n')
-        os.remove("temp_all_det_cors.txt")
+        os.remove(os.path.join(temp_dir,'temp_all_det_cors.txt'))
 
-    def sort_detection(self, temp_dir):
+    def sort_detection(self, temp_dir, output_dir):
         origin_file = temp_dir
-        output_file = "final_"+temp_dir
+        output_file = temp_dir + "_final"
 
         if not os.path.isdir(output_file):
             os.mkdir(output_file)
 
-        files = glob.glob(origin_file+'*.txt')
+        files = glob.glob(origin_file+'/*.txt')
         files.sort()
 
         for i in files:
@@ -157,6 +172,7 @@ class TextEvaluator(DatasetEvaluator):
                 outstr = outstr+',####' + rec
                 fout.writelines(outstr+'\n')
             fout.close()
+        current_path = os.getcwd()
         os.chdir(output_file)
 
         def zipdir(path, ziph):
@@ -165,14 +181,15 @@ class TextEvaluator(DatasetEvaluator):
                 for file in files:
                     ziph.write(os.path.join(root, file))
 
+        det_path = os.path.join(output_dir, 'det.zip')
         zipf = zipfile.ZipFile('../det.zip', 'w', zipfile.ZIP_DEFLATED)
         zipdir('./', zipf)
         zipf.close()
-        os.chdir("../")
+        os.chdir(current_path)
         # clean temp files
         shutil.rmtree(origin_file)
         shutil.rmtree(output_file)
-        return "det.zip"
+        return det_path
     
     def evaluate_with_official_code(self, result_path, gt_path):
         return text_eval_script.text_eval_main(det_file=result_path, gt_file=gt_path, is_word_spotting=self._word_spotting)
@@ -203,10 +220,14 @@ class TextEvaluator(DatasetEvaluator):
 
         self._results = OrderedDict()
         
+        if not self._text_eval_gt_path:
+            return copy.deepcopy(self._results)
         # eval text
-        temp_dir = "temp_det_results/"
+        temp_dir = os.path.join(self._output_dir, "temp_det_results")
+        if not os.path.isdir(temp_dir):
+            os.mkdir(temp_dir)
         self.to_eval_format(file_path, temp_dir, self._text_eval_confidence)
-        result_path = self.sort_detection(temp_dir)
+        result_path = self.sort_detection(temp_dir, self._output_dir)
         text_result = self.evaluate_with_official_code(result_path, self._text_eval_gt_path)
         os.remove(result_path)
 
@@ -220,72 +241,81 @@ class TextEvaluator(DatasetEvaluator):
         return copy.deepcopy(self._results)
 
 
-def instances_to_coco_json(instances, img_id):
-    num_instances = len(instances)
-    if num_instances == 0:
-        return []
+    def instances_to_coco_json(self, instances, img_id):
+        num_instances = len(instances)
+        if num_instances == 0:
+            return []
 
-    scores = instances.scores.tolist()
-    beziers = instances.beziers.numpy()
-    recs = instances.recs.numpy()
-
-    results = []
-    for bezier, rec, score in zip(beziers, recs, scores):
-        # convert beziers to polygons
-        poly = bezier_to_polygon(bezier)
-        s = decode(rec)
-        result = {
-            "image_id": img_id,
-            "category_id": 1,
-            "polys": poly,
-            "rec": s,
-            "score": score,
-        }
-        results.append(result)
-    return results
-
-
-def bezier_to_polygon(bezier):
-    u = np.linspace(0, 1, 20)
-    bezier = bezier.reshape(2, 4, 2).transpose(0, 2, 1).reshape(4, 4)
-    points = np.outer((1 - u) ** 3, bezier[:, 0]) \
-        + np.outer(3 * u * ((1 - u) ** 2), bezier[:, 1]) \
-        + np.outer(3 * (u ** 2) * (1 - u), bezier[:, 2]) \
-        + np.outer(u ** 3, bezier[:, 3])
-    
-    # convert points to polygon
-    points = np.concatenate((points[:, :2], points[:, 2:]), axis=0)
-    return points.tolist()
-
-
-CTLABELS = [' ','!','"','#','$','%','&','\'','(',')','*','+',',','-','.','/','0','1','2','3','4','5','6','7','8','9',':',';','<','=','>','?','@','A','B','C','D','E','F','G','H','I','J','K','L','M','N','O','P','Q','R','S','T','U','V','W','X','Y','Z','[','\\',']','^','_','`','a','b','c','d','e','f','g','h','i','j','k','l','m','n','o','p','q','r','s','t','u','v','w','x','y','z','{','|','}','~']
-
-
-def ctc_decode(rec):
-    # ctc decoding
-    last_char = False
-    s = ''
-    for c in rec:
-        c = int(c)
-        if c < 95:
-            if last_char != c:
-                s += CTLABELS[c]
-                last_char = c
-        elif c == 95:
-            s += u'口'
+        scores = instances.scores.tolist()
+        if hasattr(instances, 'beziers'):
+            beziers = instances.beziers.numpy()
         else:
-            last_char = False
-    return s
+            beziers = instances.pred_boxes.tensor.numpy()
+        recs = instances.recs.numpy()
+
+        results = []
+        for bezier, rec, score in zip(beziers, recs, scores):
+            if hasattr(instances, 'beziers'):
+                # convert beziers to polygons
+                poly = self.bezier_to_polygon(bezier)
+            else:
+                b = bezier.astype('float')
+                poly = [[b[0], b[1]], [b[0], b[3]], [b[2],b[3]], [b[2], b[1]]]
+            s = self.decode(rec)
+            result = {
+                "image_id": img_id,
+                "category_id": 1,
+                "polys": poly,
+                "rec": s,
+                "score": score,
+            }
+            results.append(result)
+        return results
 
 
-def decode(rec):
-    s = ''
-    for c in rec:
-        c = int(c)
-        if c < 95:
-            s += CTLABELS[c]
-        elif c == 95:
-            s += u'口'
+    def bezier_to_polygon(self, bezier):
+        u = np.linspace(0, 1, 20)
+        bezier = bezier.reshape(2, 4, 2).transpose(0, 2, 1).reshape(4, 4)
+        points = np.outer((1 - u) ** 3, bezier[:, 0]) \
+            + np.outer(3 * u * ((1 - u) ** 2), bezier[:, 1]) \
+            + np.outer(3 * (u ** 2) * (1 - u), bezier[:, 2]) \
+            + np.outer(u ** 3, bezier[:, 3])
+        
+        # convert points to polygon
+        points = np.concatenate((points[:, :2], points[:, 2:]), axis=0)
+        return points.tolist()
 
-    return s
-            
+    def ctc_decode(self, rec):
+        # ctc decoding
+        last_char = False
+        s = ''
+        for c in rec:
+            c = int(c)
+            if c < self.voc_size - 1:
+                if last_char != c:
+                    if self.voc_size == 96:
+                        s += self.CTLABELS[c]
+                        last_char = c
+                    else:
+                        s += str(chr(self.CTLABELS[c]))
+                        last_char = c
+            elif c == self.voc_size -1:
+                s += u'口'
+            else:
+                last_char = False
+        return s
+    
+    
+    def decode(self, rec):
+        s = ''
+        for c in rec:
+            c = int(c)
+            if c < self.voc_size - 1:
+                if self.voc_size ==96:
+                    s += self.CTLABELS[c]
+                else:
+                    s += str(chr(self.CTLABELS[c]))
+            elif c == self.voc_size -1:
+                s += u'口'
+    
+        return s
